+++ src/components/TodoEdit.tsx
@@ -12,8 +12,9 @@ import { todayDate } from "~/lib/lib"
 import {
   ClientSubtask,
   ClientTodo,
-  useGlobalContext,
-} from "../GlobalContext/store"
+  TodoListMode,
+  useTodoList,
+} from "../GlobalContext/todo-list"
 import Icon from "./Icon"
 
 interface Props {
@@ -21,7 +22,7 @@ interface Props {
 }
 
 export default function TodoEdit(props: Props) {
-  const global = useGlobalContext()
+  const todoList = useTodoList()
   const [title, setTitle] = createSignal(props.todo.title)
   const [note, setNote] = createSignal(props.todo.note)
   const [dueDate, setDueDate] = createSignal(props.todo.dueDate)
@@ -31,20 +32,20 @@ export default function TodoEdit(props: Props) {
   const [starred, setStarred] = createSignal(props.todo.starred)
 
   onMount(() => {
-    global.setClickTimeStamp(0)
+    todoList.setClickTimeStamp(0)
   })
 
   onCleanup(() => {
     // REMOVE
     if (title() === "") {
-      global.todosState.removeTodo(props.todo.key)
+      todoList.todosState.removeTodo(props.todo.key)
       return
     }
 
-    if ("subtasks" in global.flatTasks()[global.focusedTodoIndex()]) {
+    if ("subtasks" in todoList.flatTasks()[todoList.focusedTodoIndex()]) {
       // UPDATE TASK
       batch(() => {
-        global.todosState.updateTodo(props.todo.key, (p) => ({
+        todoList.todosState.updateTodo(props.todo.key, (p) => ({
           ...p,
           title: title(),
           note: note(),
@@ -52,17 +53,17 @@ export default function TodoEdit(props: Props) {
           starred: starred(),
           dueDate: showCalendar() && !dueDate() ? todayDate() : dueDate(),
         }))
-        global.setEditingTodo(false)
+        todoList.setTodoListMode(TodoListMode.Default)
       })
       return
     }
 
     // UPDATE SUBTASK
     batch(() => {
-      global.todosState.updateSubtask(
+      todoList.todosState.updateSubtask(
         // TODO: not sure how to avoid ts-ignore..
         // @ts-ignore
-        global.flatTasks()[global.focusedTodoIndex()].parent.key,
+        todoList.flatTasks()[todoList.focusedTodoIndex()].parent.key,
         props.todo.key,
         // somehow id, key and parent need to be on the type
         // but they have to be derived, don't know how..
@@ -76,7 +77,7 @@ export default function TodoEdit(props: Props) {
           done: false,
         })
       )
-      global.setEditingTodo(false)
+      todoList.setTodoListMode(TodoListMode.Default)
     })
   })
 
@@ -84,19 +85,15 @@ export default function TodoEdit(props: Props) {
     noteRef!: HTMLInputElement,
     datePickerRef!: HTMLInputElement
 
-  createEffect(() => {
-    if (global.editNoteInTodo()) {
-      autofocus(noteRef)
+  const [editNoteInTodo, setEditNoteInTodo] = createSignal(false)
 
-      createShortcut(["ArrowUp"], () => {
-        global.setEditNoteInTodo(false)
-      })
+  createEffect(() => {
+    if (editNoteInTodo()) {
+      autofocus(noteRef)
+      createShortcut(["ArrowUp"], () => setEditNoteInTodo(false))
     } else {
       autofocus(titleRef)
-
-      createShortcut(["ArrowDown"], () => {
-        global.setEditNoteInTodo(true)
-      })
+      createShortcut(["ArrowDown"], () => setEditNoteInTodo(true))
     }
   })
