+++ src/pages/App.tsx
@@ -1,78 +1,64 @@
-import { createShortcut } from "@solid-primitives/keyboard"
-import { Show, batch } from "solid-js"
-import { PageType, useGlobalContext } from "~/GlobalContext/store"
+import { Show, Signal, createContext, createSignal, useContext } from "solid-js"
+import { createSettingsState } from "~/GlobalContext/settings"
 import Help from "~/components/Help"
 import Modal from "~/components/Modal"
 import Settings from "~/components/Settings"
 import Sidebar from "~/components/Sidebar"
 import TodoList from "~/components/TodoList"
-import { todayDate } from "~/lib/lib"
+import { createShortcuts } from "~/lib/primitives"
+
+export enum PageType {
+  All = "All",
+  Today = "Today",
+  Done = "Done",
+  Starred = "Starred",
+}
+
+const ActivePageContext = createContext<Signal<PageType>>()
+export const useActivePage = () => useContext(ActivePageContext)!
 
 export default function App() {
-  const global = useGlobalContext()
+  const settingsState = createSettingsState()
 
-  createShortcut(
-    ["Control", "1"],
-    () => {
-      batch(() => {
-        global.setActivePage(PageType.All)
-        global.setFocusedTodoIndex(0)
-      })
-    },
-    { preventDefault: false }
-  )
+  const [showHelp, setShowHelp] = createSignal(false)
+  const [showSettings, setShowSettings] = createSignal(false)
+  const [activePage, setActivePage] = createSignal(PageType.All)
 
-  createShortcut(
-    ["Control", "2"],
-    () => {
-      batch(() => {
-        global.setActivePage(PageType.Today)
-        global.setFocusedTodoIndex(0)
-      })
+  createShortcuts({
+    "Control+1"() {
+      setActivePage(PageType.All)
     },
-    { preventDefault: false }
-  )
-
-  createShortcut(
-    ["Control", "3"],
-    () => {
-      batch(() => {
-        global.setActivePage(PageType.Starred)
-        global.setFocusedTodoIndex(0)
-      })
+    "Control+2"() {
+      setActivePage(PageType.Today)
     },
-    { preventDefault: false }
-  )
-
-  createShortcut(
-    ["Control", "4"],
-    () => {
-      batch(() => {
-        global.setActivePage(PageType.Done)
-        global.setFocusedTodoIndex(0)
-      })
+    "Control+3"() {
+      setActivePage(PageType.Starred)
     },
-    { preventDefault: false }
-  )
+    "Control+4"() {
+      setActivePage(PageType.Done)
+    },
+  })
 
   return (
     <div class="flex min-h-screen  bg-gray-100 dark:bg-stone-900">
-      <Sidebar />
-      <TodoList />
-      <Show when={global.showHelp()}>
-        <Modal
-          title="Help"
-          onClose={() => global.setShowHelp(false)}
-          children={<Help />}
-        />
-      </Show>
-      <Show when={global.showSettings()}>
-        <Modal
-          title="Settings"
-          onClose={() => global.setShowSettings(false)}
-          children={<Settings />}
-        />
-      </Show>
+      <ActivePageContext.Provider value={[activePage, setActivePage]}>
+        <Sidebar />
+        <TodoList />
+        <Show when={showHelp()}>
+          <Modal
+            title="Help"
+            onClose={() => setShowHelp(false)}
+            children={<Help />}
+          />
+        </Show>
+        <Show when={showSettings()}>
+          <Modal
+            title="Settings"
+            onClose={() => setShowSettings(false)}
+            children={<Settings />}
+          />
+        </Show>
+      </ActivePageContext.Provider>
     </div>
   )
 }
